requirements.txt
fastapi==0.110.0
mangum==0.17.0
openai==1.14.2
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
uvicorn==0.29.0

